Apply PEP8 standards
Linting with pylint so the PEP8 standards are uphelp throughout the codebase. Some of the pylint features had to be disabled in certain files, because they did not make sense in the given context. For instance, the objects property on a Model class was not recognized as a property, simply because it is a dynamically attached property